lingo-stage1-audio-only/lingo-v2/agent/requirements.txt
livekit-agents==1.4.6
livekit-plugins-deepgram==1.4.6
deepgram-sdk>=3.0.0
python-dotenv==1.0.1
websockets==13.1
aiohttp==3.10.0